Check projector active state using a specialized event publisher
This simplifies the projector logic as all projector state tracking is
now only checked by the `ActiveProjectorsEventPublisher`. Replay now
uses the normal `EventPublisher` so the logic to determine if an event
is allowed to be processed by a projector can be simplified.
