+++ lib/sequent/core/core.rb
@@ -19,5 +19,6 @@ require_relative 'command_record'
 require_relative 'workflow'
 require_relative 'random_uuid_generator'
 require_relative 'event_publisher'
+require_relative 'active_projectors_event_publisher'
 require_relative 'aggregate_snapshotter'
 require_relative 'projectors'
