+++ lib/sequent/core/event_publisher.rb
@@ -3,9 +3,6 @@
 module Sequent
   module Core
     class ProjectorMigrationError < RuntimeError; end
-    class UnknownActiveProjectorError < ProjectorMigrationError; end
-    class ReplayingProjectorMismatchError < ProjectorMigrationError; end
-    class DifferentProjectorVersionIsActiveError < ProjectorMigrationError; end
 
     #
     # EventPublisher ensures that, for every thread, events will be published
@@ -35,15 +32,6 @@ module Sequent
       def publish_events(events)
         return if configuration.disable_event_handlers
 
-        ensure_no_unknown_active_projectors!
-
-        events.each { |event| events_queue.push(event) }
-        process_events
-      end
-
-      def replay_events(events)
-        ensure_only_replaying_projectors_subscribed!
-
         events.each { |event| events_queue.push(event) }
         process_events
       end
@@ -68,7 +56,7 @@ module Sequent
         Sequent.logger.debug("[EventPublisher] Publishing event #{event.class}") if Sequent.logger.debug?
 
         configuration.event_handlers.each do |handler|
-          handler.handle_message event
+          handle_message(handler, event)
         rescue ProjectorMigrationError
           raise
         rescue StandardError
@@ -76,41 +64,12 @@ module Sequent
         end
       end
 
+      def handle_message(handler, event)
+        handler.handle_message(event)
+      end
+
       def configuration
         Sequent.configuration
-      end
-
-      def ensure_no_unknown_active_projectors!
-        return unless Sequent.configuration.enable_projector_states
-
-        expected_version = Sequent.migrations_class&.version
-        return if expected_version.nil?
-
-        registered_projectors = Migratable.projectors.to_set(&:name)
-        active_projectors = Projectors.projector_states
-          .values
-          .select { |s| s.active_version == expected_version }
-          .to_set(&:name)
-        unknown_active_projectors = active_projectors - registered_projectors
-        if unknown_active_projectors.present?
-          fail UnknownActiveProjectorError,
-               "cannot publish event when unknown projectors are active #{unknown_active_projectors}"
-        end
-      end
-
-      def ensure_only_replaying_projectors_subscribed!
-        return unless Sequent.configuration.enable_projector_states
-        return unless Sequent.migrations_class
-
-        registered_projectors = Migratable.projectors.to_set(&:name)
-        replaying_projectors = Projectors.projector_states
-          .values
-          .select { |state| state.replaying? || state.activating? }
-          .to_set(&:name)
-        if registered_projectors != replaying_projectors
-          fail ReplayingProjectorMismatchError,
-               "cannot replay event when different projectors are replaying #{replaying_projectors}"
-        end
       end
     end
   end
