+++ spec/lib/sequent/core/projector_spec.rb
@@ -33,77 +33,6 @@ describe Sequent::Core::Projector do
     end.to raise_error(Sequent::Core::Projector::NotManagedByThisProjector)
   end
 
-  context 'projector activation' do
-    def set_lock_timeout = ActiveRecord::Base.connection.execute("SET lock_timeout TO '10ms'")
-    def update_projector_states = Sequent::Core::Projectors.register_active_projectors!([Sequent::Core::Projector], 1)
-    def read_projector_states = Sequent::Core::Projectors.projector_states
-
-    before do
-      Sequent.configuration.enable_projector_states = true
-      Sequent.activate_current_configuration!
-    end
-    after do
-      Sequent.configuration.enable_projector_states = false
-    end
-
-    it 'waits for project state updates before reading' do
-      reader_lock_attempted = Queue.new
-      writer_locked = Queue.new
-      writer = Thread.new do
-        ActiveRecord::Base.transaction do
-          update_projector_states
-          writer_locked << 'locked states for reading'
-
-          reader_lock_attempted.deq
-        end
-      end
-
-      reader = Thread.new do
-        ActiveRecord::Base.transaction do
-          writer_locked.deq
-          set_lock_timeout
-          read_projector_states
-          false
-        rescue ActiveRecord::LockWaitTimeout
-          true
-        end
-      end
-
-      expect(reader.join(0.5).value).to be(true)
-    ensure
-      reader_lock_attempted << 'reader finished'
-      writer.join(0.5).value
-    end
-
-    it 'waits for readers to when updating the projector states table' do
-      writer_lock_attempted = Queue.new
-      reader_locked = Queue.new
-      reader = Thread.new do
-        ActiveRecord::Base.transaction do
-          read_projector_states
-          reader_locked << 'locked states for reading'
-
-          writer_lock_attempted.deq
-        end
-      end
-
-      writer = Thread.new do
-        ActiveRecord::Base.transaction do
-          reader_locked.deq
-          set_lock_timeout
-          update_projector_states
-        rescue ActiveRecord::LockWaitTimeout
-          true
-        end
-      end
-
-      expect(writer.join(0.5).value).to be(true)
-    ensure
-      writer_lock_attempted << 'writer finished'
-      reader.join(0.5).value
-    end
-  end
-
   context 'forward methods' do
     let(:persistor) { double('persistor') }
     let(:record) { double('record') }
